+++ tests/integration/common.rs
@@ -15,7 +15,7 @@ use miden_client::{
     },
     sync::SyncSummary,
     transactions::{
-        transaction_request::{TransactionRequest, TransactionTemplate},
+        request::{TransactionRequest, TransactionTemplate},
         DataStoreError, TransactionExecutorError,
     },
     Client, ClientError,
@@ -243,7 +243,7 @@ pub async fn mint_note(
 
     // Check that note is committed and return it
     println!("Fetching Committed Notes...");
-    let note_id = tx_request.expected_output_notes()[0].id();
+    let note_id = tx_request.expected_output_notes().next().unwrap().id();
     let note = client.get_input_note(note_id).unwrap();
     note.try_into().unwrap()
 }